Introduce Watchlist feature.

+++ network/src/main/java/de/r4md4c/gamedealz/network/service/IsThereAnyDealService.kt
@@ -75,7 +75,7 @@ internal interface IsThereAnyDealService {
         @Query("region") region: String? = null,
         @Query("country") country: String? = null,
         @Query("shops") shops: String? = null,
-        @Query("added") added: Int? = null
+        @Query("added") added: Long? = null
     )
             : Deferred<DataWrapper<PlainPriceList>>
